+++ R/teal.transform-package.R
@@ -4,5 +4,5 @@
 
 #' @import shiny
 #' @importFrom magrittr %>%
-#' @importFrom formatable var_relabel var_labels
+#' @importFrom formatters var_relabel var_labels
 NULL