+++ NEWS.md
@@ -6,4 +6,4 @@
 * Added new argument `ordered` for `select_spec()` to flag whether order of the selection should be tracked.
 
 ### Misc
-* The `get_relabel_call` function now returns `formatable::var_relabel` instead of `teal.data::var_relabel`.
+* The `get_relabel_call` function now returns `formatters::var_relabel` instead of `teal.data::var_relabel`.